# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Odoo Cafe POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [payment-methods]
      summary: List payment methods
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
    post:
      operationId: createPaymentMethod
      tags: [payment-methods]
      summary: Create payment method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
  /payment-methods/{id}:
    patch:
      operationId: updatePaymentMethod
      tags: [payment-methods]
      summary: Update payment method
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
    delete:
      operationId: deletePaymentMethod
      tags: [payment-methods]
      summary: Delete payment method
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── FLOORS ───────────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: sessionId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [Draft, Paid, Cancelled]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create draft order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order (lines, customer, discount)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete draft order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
  /orders/{id}/pay:
    post:
      operationId: payOrder
      tags: [orders]
      summary: Pay for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayOrderInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/send-kitchen:
    post:
      operationId: sendOrderToKitchen
      tags: [orders]
      summary: Send order to kitchen display
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── SESSIONS ─────────────────────────────────────────────────────────────────
...
  /kds:
    get:
      operationId: listKdsTickets
      tags: [kds]
      summary: List KDS tickets
      parameters:
        - name: stage
          in: query
          schema:
            type: string
            enum: [ToCook, Preparing, Completed]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KdsTicket"
  /kds/{id}/advance:
    post:
      operationId: advanceKdsTicket
      tags: [kds]
      summary: Advance ticket to next stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KdsTicket"
  /kds/{id}/items/{itemIndex}/toggle:
    post:
      operationId: toggleKdsItem
      tags: [kds]
      summary: Toggle item done status in ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemIndex
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KdsTicket"

  # ── REPORTS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PaymentMethod:
      type: object
      required: [id, name, type, active]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [Cash, Card, UPI]
        upiId:
          type: ["string", "null"]
        active:
          type: boolean
    PaymentMethodInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [Cash, Card, UPI]
        upiId:
          type: string
        active:
          type: boolean
    PaymentMethodUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [Cash, Card, UPI]
        upiId:
          type: ["string", "null"]
        active:
          type: boolean

    # ── Floor ─────────────────────────────────────────────────────────────────
...
    CartLine:
      type: object
      required: [productId, qty, unitPrice]
      properties:
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        qty:
          type: number
        unitPrice:
          type: number
        lineTotal:
          type: number
        productDiscount:
          type: ["number", "null"]
        productDiscountLabel:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, number, lines, subtotal, tax, discountTotal, total, status, createdAt]
      properties:
        id:
          type: integer
        number:
          type: string
        tableId:
          type: ["integer", "null"]
        tableNumber:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/CartLine"
        subtotal:
          type: number
        tax:
          type: number
        discountTotal:
          type: number
        discountLabel:
          type: ["string", "null"]
        total:
          type: number
        status:
          type: string
          enum: [Draft, Paid, Cancelled]
        paymentMethodId:
          type: ["integer", "null"]
        paymentRef:
          type: ["string", "null"]
        amountPaid:
          type: ["number", "null"]
        sentToKitchen:
          type: boolean
        employeeId:
          type: ["integer", "null"]
        sessionId:
          type: ["integer", "null"]
        createdAt:
          type: string
    OrderInput:
      type: object
      properties:
        tableId:
          type: integer
        sessionId:
          type: integer
    OrderUpdate:
      type: object
      properties:
        tableId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        lines:
          type: array
          items:
            type: object
            required: [productId, qty, unitPrice]
            properties:
              productId:
                type: integer
              qty:
                type: number
              unitPrice:
                type: number
        couponCode:
          type: ["string", "null"]
        discountKind:
          type: ["string", "null"]
          enum: [percent, amount, null]
        discountValue:
          type: ["number", "null"]
        discountLabel:
          type: ["string", "null"]
    PayOrderInput:
      type: object
      required: [paymentMethodId]
      properties:
        paymentMethodId:
          type: integer
        amountPaid:
          type: number
        paymentRef:
          type: string

    # ── POS Session ───────────────────────────────────────────────────────────
...
    KdsItem:
      type: object
      required: [productId, qty, done]
      properties:
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        qty:
          type: number
        done:
          type: boolean
    KdsTicket:
      type: object
      required: [id, orderId, orderNumber, items, stage, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        orderNumber:
          type: string
        tableNumber:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/KdsItem"
        stage:
          type: string
          enum: [ToCook, Preparing, Completed]
        createdAt:
          type: string

    # ── Reports ───────────────────────────────────────────────────────────────
...